twitter_analysis_spark/twitter-analysis/requirements.txt
kafka-python==0.9.4
oauthlib==1.0.3
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
tweepy==3.3.0
configparser==3.3.0.post2
numpy==1.13.3